Removed QueryAugmenterService, moved logic to ExecutableDocument And renamed to AbstractExecutableDocument
