+++ layers/API/packages/api-graphql/src/DataStructureFormatters/GraphQLDataStructureFormatter.php
@@ -9,10 +9,10 @@ use PoPAPI\GraphQLAPI\Module;
 use PoPAPI\GraphQLAPI\ModuleConfiguration;
 use PoP\ComponentModel\App;
 use PoP\ComponentModel\Constants\Response;
+use PoP\ComponentModel\ExtendedSpec\Execution\ExecutableDocument;
 use PoP\ComponentModel\Feedback\FeedbackCategories;
 use PoP\ComponentModel\Feedback\FeedbackEntryManagerInterface;
 use PoP\ComponentModel\Feedback\Tokens;
-use PoP\GraphQLParser\ExtendedSpec\Execution\ExecutableDocument;
 use PoP\GraphQLParser\FeedbackItemProviders\GraphQLSpecErrorFeedbackItemProvider;
 use PoP\GraphQLParser\Spec\Parser\Ast\FieldInterface;
 use PoP\GraphQLParser\Spec\Parser\Ast\LeafField;
