--- layers/Engine/packages/graphql-parser/src/Facades/Query/QueryAugmenterServiceFacade.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PoP\GraphQLParser\Facades\Query;
-
-use PoP\Root\App;
-use PoP\GraphQLParser\Query\QueryAugmenterServiceInterface;
-
-class QueryAugmenterServiceFacade
-{
-    public static function getInstance(): QueryAugmenterServiceInterface
-    {
-        /**
-         * @var QueryAugmenterServiceInterface
-         */
-        $service = App::getContainer()->get(QueryAugmenterServiceInterface::class);
-        return $service;
-    }
-}
